Fix build on linux with libmagic (#182)
* Travis: install libmagic

* Fix build on linux with libmagic; fixes #181

When libmagic is installed on Linux, it is used by libmtp-ng-static.a.

And when libmtp-ng-static.a is linked to an executable, the arguments
passed to c++ must include `libmtp-ng-static.a -lmagic`. The order of
these arguments must be in this order since libmtp-ng-static.a depends
libmagic (so-called "inter-library dependency"). However, the order have
been reversed -- `-lmagic libmtp-ng-static.a`, causing errors of
"undefined reference".

This changes assures this order so that the inter-library dependency
will be correctly resolved.